refactor/projects/punch_classification/pyproject.toml
[project]
name = "punch-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"

dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.7.0",
    "pandas>=2.0.0",
    "pillow>=11.1.0",
    "scikit-learn>=1.2.0",
    "torch==2.5.0+cu124; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torch==2.5.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
name = "torch-cuda"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "torch-cuda", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'"},
    { index = "torch-cpu", marker = "sys_platform == 'arm64'"},
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.9.10",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
